Added: base shell code.
All I have done is created a basic shell template that I can extend
from. This will not be very big now but will grow in the future.

I will not reinvent the wheel and just keep it simple by sticking to the
GNU utils "cd, ls, echo, ps aux"